+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added autoLocation options for weather forcast and current weather modules.
 - Added autoTimezone option for the default clock module.
 - Danish translation for "Feels" and "Weeks"
+- Added option to split multiple day events in calendar to separate numbered events
 
 ### Updated
 - Bumped the Electron dependency to v3.0.13 to support the most recent Raspbian. [#1500](https://github.com/MichMich/MagicMirror/issues/1500)